+++ test/dummy/config/routes.rb
@@ -1,3 +1,4 @@
 Dummy::Application.routes.draw do
+  resources :articles, :only => [:create]
   resources :widgets
 end